Reconfigure
There were a few problems with the old version from 3-10. So, I modified some of the function names. I also got rid of the defaul '/' method that gives you the routes. This is because the routes json file is poorly organized and needs to fixed later. I also moved the main data file into its own folder. I also split up the main data folder into two different json files. This is so that whatever route you call works only with relevent data and not all the data.