Canned Food
I have changed a few things. The nutrition data I got for the canned items were not from the USDA, like the fruits. So, I removed USDA from the canned data. I also added a few more canned products. Lastly, I updated the docs. There was typo that needed fixed as well as an update on the stats section.