Added new items
I added 3 new canned items. I also updated all the canned items to include a serving per container. This was because the information in the nutrtion section is actually only part of whats in the product. So, the per container lets the consumer know how much is actually is in the whole container with some quick math.