Nutrition Changes
I have added one more item to the dairy. In doing this I realized that there was an issue with the nutrition. The amounts given for each nutrient group is from the serving size not the whole container. Essentially, if it says 4 servings per container, there are 4 times the amount of nutrients listed. All of the products have been inputted just fine. However, the labels still included the word 'total' for carbohydrates and fat. I have removed that to show that it is not the total amount in the package, but instead the amount in that serving.