Updated docs and routes
I noticed a disparity with the docs.json file, which was not up to date. So, I fixed it and added all available routes thus far. Also, the default '/' route now returns documentation on the routes. This way, you can easily pull up the routes by just using the api. I also created a new '/all' route. This route will return all of the item data, which was previously a part of the default '/' route.