Added Temperature Regulation
I have implemented temperature regulation values to each product. There is the 'hot' value, this should be used for things that must be kept hot or are sold as being hot. This can be things like fresh chicken. The next one is 'cold', this value should be used for cold/refridgerated things. Things in this catetgory include dairy products like milk and cheese. The final temperature value is 'frozen', this is for things that need to be kept frozen. This can be things like frozen dinners. Even though there are no frozen items yet, I thought it would be a good idea to have an option for frozen. I also updated the Readme a little bit to include the additions.