Updated Controller
With the addition of the canned seafood there was a problem of some products having multiple categories. To solve this, I added an array with multiple categories. I also added to the search controller to also check for the possibility of an array with multiple values instead of just one value.