Data Layout Change
There was an issue with specific searches that wouldn't not show the name of the products, just the data. I made the entire data file into an array with an object for each product. I have moved the name into its own data value. I thought this new layout would be a conflict to the route controllers. However, the way I wrote them allows them to work with both objects and arrays. If there are any issues that I have overlooked with this change, please leave an issue.