Added Stats Route
There is now a route for stats on the data. The first thing I have added to this route is the number of items for each category. This will make adding items far easier, so there can be more of a distribution of data rather than a bunch of one category. I also noticed a few mistakes, so I took care of those.