Mistake
I was testing the stats route and in doing so, I have found an error in the number of categories. I do not know how this error happened. Due to the large amount of categories being added, it was bound to happen at some point. It was listed as 70 categories, when in reality there were 76 categories. This shows the importance of expanding the stats route to include brands, and products.