Stats Route Change
With the new updates, there was a need to rework the stats route. What I have added doesn't really change the stats route a whole lot, but it does fix a few structural things. There are now country parameters you can attach to the end of the stats route. This will give information based on each country. I have also created a master list and master stats template. These include information on all items of all countries. They are located in the main data folder and are strictly for reference. Quite a few of the brands and categoreis may not work with every country. The master stats template will be used for future stats including all products. The final things changed was the food naming. There was a problem with putting countries in a search url. Spacing and capitalization may be difficult and teadious. So, it has been simplified to just use things like 'US' instead of United States and 'JP' instead of Japan.