Country Changes
The country system worked relatively well. However, many products are multinational. This presents an issue with adding new products. So, all data has been combined into a single file. However, there has been the addition of a country parameter. Now, products will have countries in which they are sold in. This is to make it easier to add new products, while at the same time also having the ability to search for items that may be specific to certain countries.